src/plant-wall-control/requirements.txt
Flask==2.0.1
RPi.GPIO==0.7.0
Adafruit-DHT==1.4.0
requests==2.26.0
numpy==1.21.2
Pillow==8.4.0
flask-cors==3.0.10